+++ plugins/scaffolder-actions/scaffolder-backend-module-utils/src/actions/fs/parseFile.test.ts
@@ -40,7 +40,7 @@ describe('roadiehq:utils:fs:parse', () => {
         ...mockContext,
         input: {} as any,
       }),
-    ).rejects.toThrow(/"path" argument must/);
+    ).rejects.toThrow(/argument must be of type string/);
   });
   it('should pasrs text data by default', async () => {
     mock({