+++ plugins/scaffolder-actions/scaffolder-backend-module-utils/src/actions/jsonata/yaml.ts
@@ -17,14 +17,15 @@ import { createTemplateAction } from '@backstage/plugin-scaffolder-backend';
 import jsonata from 'jsonata';
 import { resolveSafeChildPath } from '@backstage/backend-common';
 import fs from 'fs-extra';
-import yaml from 'js-yaml';
-import { supportedDumpOptions, yamlOptionsSchema } from '../../types';
+import YAML from 'yaml';
+import { stringifyOptions, yamlOptionsSchema } from '../../types';
 
 export function createYamlJSONataTransformAction() {
   return createTemplateAction<{
     path: string;
     expression: string;
-    options?: supportedDumpOptions;
+    options?: stringifyOptions;
+    loadAll?: boolean;
     as?: 'string' | 'object';
   }>({
     id: 'roadiehq:utils:jsonata:yaml:transform',
@@ -45,6 +46,12 @@ export function createYamlJSONataTransformAction() {
             title: 'Expression',
             description: 'JSONata expression to perform on the input',
             type: 'string',
+          },
+          loadAll: {
+            title: 'Load All',
+            description:
+              'Use this if the yaml source file contains multiple yaml objects',
+            type: 'boolean',
           },
           as: {
             title: 'Desired Result Type',
@@ -72,14 +79,20 @@ export function createYamlJSONataTransformAction() {
       if (ctx.input.as === 'object') {
         resultHandler = rz => rz;
       } else {
-        resultHandler = rz => yaml.dump(rz, ctx.input.options);
+        resultHandler = rz => YAML.stringify(rz, ctx.input.options);
       }
       const sourceFilepath = resolveSafeChildPath(
         ctx.workspacePath,
         ctx.input.path,
       );
-
-      const data = yaml.load(fs.readFileSync(sourceFilepath).toString());
+      let data;
+      if (ctx.input.loadAll) {
+        data = YAML.parseAllDocuments(
+          fs.readFileSync(sourceFilepath).toString(),
+        ).map(doc => doc.toJSON());
+      } else {
+        data = YAML.parse(fs.readFileSync(sourceFilepath).toString());
+      }
       const expression = jsonata(ctx.input.expression);
       const result = expression.evaluate(data);